+++ doc/changes.rst
@@ -16,6 +16,7 @@ New in 0.27
 
 * :func:`concatenate`: concatenate multiple :class:`NDVar` objects to form a
   new dimension.
+* :meth:`NDVar.ols`: regress on a dimension.
 * New functions:
 
   - :func:`psd_welch`